requirements.txt
discord.py>=2.4,<3.0
Flask>=3.0,<4.0
Pillow>=10.0,<12.0
ImageHash>=4.3,<5.0
python-dotenv>=1.0,<2.0
psutil>=5.9,<8.0
